add:product decorate
