+++ rails/app/decorators/application_decorator.rb
@@ -1,0 +1,2 @@
+class ApplicationDecorator < Draper::Decorator
+end
