+++ rails/app/views/shared/_product_review_form.html.slim
@@ -10,7 +10,7 @@
         .fields
           =f.label :rate, class: "four wide field"
           .ten.wide.field
-            .ui.star.rating#rating
+            .ui.yellow.star.rating#rating
             =f.hidden_field(:rate)
         .fields
           =f.label :content, "レビュー内容", class: "four wide field"
