+++ rails/spec/decorators/product_decorator_spec.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+RSpec.describe ProductDecorator do
+end
